Start restructuring and auditing package.json
