+++ package.json
@@ -32,17 +32,14 @@
     "clean-webpack-plugin": "^4.0.0",
     "constants-browserify": "^1.0.0",
     "copy-webpack-plugin": "^11.0.0",
-    "core-js": "^3.32.1",
     "crypto-browserify": "^3.12.0",
     "css-loader": "^6.8.1",
-    "dotenv": "^16.3.1",
     "events": "^3.3.0",
     "html-webpack-plugin": "^5.5.3",
     "http-message-signatures": "^1.0.4",
     "httpbis-digest-headers": "^1.0.0",
     "path-browserify": "^1.0.1",
     "process": "^0.11.10",
-    "prop-types": "^15.8.1",
     "querystring-es3": "^0.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -68,15 +65,11 @@
     "@testing-library/jest-dom": "^6.1.3",
     "@testing-library/react": "^14.0.0",
     "@types/chrome": "^0.0.244",
-    "@types/inboxsdk": "^2.0.11",
     "@types/jest": "^29.5.5",
-    "@types/jquery": "^3.5.18",
-    "@types/lodash": "^4.14.197",
     "@types/node": "^20.8.4",
     "@types/react": "^18.2.21",
     "@types/react-dom": "^18.2.7",
     "@types/react-router-dom": "^5.3.3",
-    "@types/redux": "^3.6.31",
     "@types/uuid": "^9.0.5",
     "@types/webextension-polyfill": "^0.10.7",
     "@types/webpack-bundle-analyzer": "^4.6.0",
