+++ src/background/messageHandlers/getSendingPaymentPointerHandler.ts
