+++ src/background/messageHandlers/getStorageData.ts
