+++ src/background/messageHandlers/isMonetizationReadyHandler.ts
