+++ src/background/messageHandlers/pausePaymentHandler.ts
