+++ src/background/messageHandlers/runPaymentHandler.ts
