+++ src/background/messageHandlers/setIncomingPointerHandler.ts
