+++ src/content/monetizationTagManager/CustomError.ts
