+++ src/content/monetizationTagManager/MonetizationTagManager.ts
