+++ src/content/monetizationTagManager/resolvePaymentEndpoint.ts
