+++ src/content/monetizationTagManager/whenDocumentReady.ts
