+++ src/popup/Popup.tsx
@@ -2,8 +2,8 @@ import './Popup.scss'
 
 import React from 'react'
 
-import { RouterProvider } from '@/components/router-provider'
-import { PopupProvider } from '@/providers/popup.provider'
+import { RouterProvider } from '@/popup/components/router-provider'
+import { PopupProvider } from '@/popup/providers/popup.provider'
 
 const Popup = () => {
   return (
