+++ src/popup/components/__tests__/code.test.tsx
