+++ src/popup/components/__tests__/slider.test.tsx
