+++ src/popup/components/icons.tsx
