+++ src/popup/components/layout/main-layout.tsx
