+++ src/popup/components/router-provider.tsx
