+++ src/popup/components/slider.tsx
