+++ src/popup/components/switch.tsx
