+++ src/popup/providers/providers.interface.ts
