--- src/types/openControl.d.ts
@@ -1,7 +1,0 @@
-declare type OpenControls = {
-  isVisible: boolean
-  handleClose: () => void
-  handleOpen: () => void
-  handleToggle: () => void
-  isOpen: boolean
-}
