--- src/types/response.d.ts
@@ -1,6 +1,0 @@
-declare type EXTResponseType = 'SUCCESS' | 'FAILED' | 'PENDING' | 'UNAUTHORIZED' | 'AUTHENTICATED'
-
-declare type EXTResponse<T = any> = {
-  type: EXTResponseType
-  data?: T
-}
